Piccola correzione esercizio 2 (da int a double)

+++ SecondExercise.java
@@ -5,12 +5,12 @@ public class SecondExercise {
         Scanner input = new Scanner(System.in);
         try {
             System.out.println("Inserisci quanti km hai percorso");
-            int kmPercorsi = Integer.parseInt(input.nextLine());
+            double kmPercorsi = Double.parseDouble(input.nextLine());
             if (kmPercorsi == 0) {
                 throw new ArithmeticException();
             }
             System.out.println("Inserisci quanti litri hai utilizzato");
-            int ltUsati = Integer.parseInt(input.nextLine());
+            double ltUsati = Double.parseDouble(input.nextLine());
             double kmAlt = kmPercorsi / ltUsati;
             System.out.println("Km/lt = " + kmAlt);
         } catch (ArithmeticException ex) {
